Move the vault root utils to the common file on AVS example
