+++ examples/awesome-vault-service/common/common.go
@@ -1,12 +1,42 @@
 package examplecommon
 
+import (
+	"slices"
+
+	"github.com/ethereum/go-ethereum/crypto"
+)
+
 type TaskInput struct {
 	Key   string
 	Value string
 }
 
-// This function computes simulates the vault setting behavior
-func VaultSet(taskIndex uint32, input TaskInput) ([32]byte, error) {
-	// TODO: complete
-	return [32]byte{0}, nil
+func ComputeVaultsRoot(vaults []TaskInput) [32]byte {
+	leaves := make([][32]byte, len(vaults))
+	for i, vault := range vaults {
+		leaves[i] = hashVault(vault)
+	}
+	for len(leaves) > 1 {
+		halfLength := (len(leaves) + 1) / 2
+		for i := range halfLength {
+			rightIdx := i*2 + 1
+			if rightIdx >= len(leaves) {
+				rightIdx = i * 2
+			}
+			leaves[i] = hashNodes(leaves[i*2], leaves[rightIdx])
+		}
+		leaves = leaves[:halfLength]
+	}
+	return leaves[0]
+}
+
+func hashVault(input TaskInput) [32]byte {
+	return crypto.Keccak256Hash([]byte(input.Key + input.Value))
+}
+
+func hashNodes(leftNode [32]byte, rightNode [32]byte) [32]byte {
+	if slices.Compare(leftNode[:], rightNode[:]) > 0 {
+		leftNode, rightNode = rightNode, leftNode
+	}
+	return crypto.Keccak256Hash(leftNode[:], rightNode[:])
 }
